Fixed zip code error (#2)
Looks like Amtrak API can't make up their mind. They mostly serialize
the zip code as a String but for an unknown zip code (why is it
unknown?) they will just use the integer type and serialize a 0. I will
handle this more elegantly in the future but this will do for now.