Added serde_debugging feature (#5)
* Added serde_debugging feature

We use to use the test feature flag but these tests are not unit tests
and therefore should not go into the mod test { } namespace. They
instead should be placed as part of an integration tests. Other users
might want this feature so instead of strictly making this a test
feature we will make it an optional feature that developers can enable
if they so like.

* Fix CI